Re-organize generated python code
Firstly decode from memoryview instead of from bytes, this it to prevent
un-necessary copying of buffers.

The behaviour of `__bytes__` and `_frombuf` has changed so that they
return the raw object without any prepended length field. These are
suitable for one-off encode/decode.

A base class is added for all classes, which provides the following
new methods:
1. `_len_wrap()` produce bytes with a length field prepended
2. `_len_unwrap()` parses bytes with a length field prepended, returns
  that length, as well as the parsed object

It also provides type prototypes for all the other methods so that
`_Xpdt` can be used as a placeholder for any struct included in the
module.

All classes also define the following new methods:
1. `_read()` which is either an alias for `__bytes__` for fixed-length
   structs,  or an alias for `_len_wrap()` for variable-length structs.
2. `_write()` which is either an alias for `_frombuf` for fixed-length
   structs, or an alias for `_len_unwrap()` for variable-length structs.
3. `_read_many()` which returns a generator for reading many items from
   a buffer. For variable-length structs, it will assume the file has
   been written with lengths prepended as if written with `_write()`.
   For fixed-length structs it will read the file as though it were an
   array.
4. `_from_file()` which memory-maps a file and calls `_read_many()`

Several micro-optimizations were also made to the generated code.

The purpose of these changes is to prepare for supporting enums which
will have a different header prepended - one which includes discriminant
as well as, possibly, the length of the object.