Generate C wrappers
The C implementation is not prefixed with a namespace, making it useful
only for some implementations. It has the benefit of having all being in
a single file and having a wide variety of functionality and relatively
terse names for things. Of particular concern is that structs are named
without any namespace prefix. This limits the direct use of this header
in some applications.

Add new backends which generate a C file and a header which expose only
functions, only for enum based event output, and with prefixed names for
the functions. This allows multiple namespaces with conflicting names
for things to coincide peacefully in the same binary object.