Fix C code generation for non-registry modules
If no registry is provided, the module is compiled in fixed-mode, there
are no discriminants so we cannot write a wrapped struct.

In this case the caller must make sure to only write structs of the same
type to a given stream or there will be no way for the reader to
distinguish them.